+++ src/mmt-api/src/main/info/kwarc/mmt/api/objects/Matcher.scala
@@ -99,6 +99,7 @@ class Matcher(controller: Controller, rules: RuleSet) extends Logger {
          case _ =>
             return None
       }
+      def materialize(cont : Context, tm : Term, expandDefs : Boolean, parent : Option[MPath]) = controller.simplifier.materialize(cont,tm,expandDefs,parent)
    }
 
    /**
