+++ src/mmt-odk/src/info/kwarc/mmt/odk/Plugin.scala
@@ -90,3 +90,38 @@ object IntegerSubtype extends SubtypingRule {
     case _ => None
   }
 }
+
+object LFX {
+  class LFRecSymbol(name:String) {
+    val path = (DPath(URI.http colon "gl.mathhub.info") / "MMT" / "LFX" / "Records") ? "Symbols" ? name
+    val term = OMS(path)
+  }
+
+  case class RecordBody(self: Option[LocalName], fields: List[OML]) {
+    /** names of all fields */
+    def names = fields.map(_.name)
+    /** checks for duplicate names */
+    def hasDuplicates = utils.hasDuplicates(names)
+    /** retrieve a field for a given name */
+    def get(l: LocalName) = fields.find(_.name == l)
+  }
+
+  /** unifies record terms and types; the empty record is OMA(this.term,Nil), not this.term */
+  class RecordLike(n: String) extends LFRecSymbol(n) {
+    // there may not be an apply method that takes a context instead of a List[OML]
+    def apply(v:OML*): Term = apply(None, v.toList)
+    def apply(self: Option[LocalName], fields: List[OML]): Term = {
+      self match {
+        case None => OMA(this.term, fields)
+        case Some(l) => OMBINDC(this.term, Context(VarDecl(l)), fields)
+      }
+    }
+    def unapply(t : Term) : Option[RecordBody] = t match {
+      case OMA(this.term, OMLList(fs)) => Some(RecordBody(None, fs))
+      case OMBINDC(this.term, Context(VarDecl(n, None, None, None, _), rest@_*), OMLList(fs)) if rest.isEmpty => Some(RecordBody(Some(n), fs))
+      case _ => None
+    }
+  }
+
+  object RecExp extends RecordLike("Recexp")
+}
