+++ src/mmt-pvs/src/info/kwarc/mmt/pvs/Plugin.scala
@@ -228,7 +228,7 @@ import PVSTheory._
 
 object SetsubRule extends ComputationRule(PVSTheory.expr.path) {
   def apply(check: CheckingCallback)(tm: Term, covered: Boolean)(implicit stack: Stack, history: History): Option[Term] = tm match {
-    case expr(setsub(tp,prop)) => Some(info.kwarc.mmt.LFX.Subtyping.predsubtp(expr(tp),proof("internal_judgment",
+    case expr(setsub(tp,prop)) => Some(LFX.predsubtp(expr(tp),proof("internal_judgment",
       Lambda(doName,expr(tp),pvsapply(prop,OMV(doName),expr(tp),bool.term)._1))))
     case _ => None
   }
